% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_if_not.r
\name{install_n_require}
\alias{install_n_require}
\title{Install if not and require package}
\usage{
install_n_require(packagenames)
}
\arguments{
\item{packagenames}{vector of package names that need to require for installation}
}
\value{
`install_n_require()` returns an error message when error
}
\description{
This is a basic interface of installing required package automatically when need
The output of `install_if_not()` will be an error of corresponding when installation failed, otherwise quiet
}
\examples{
install_n_require(packagenames = c('ggplot2','ggfortify'))

}
